+++ packages/spacecat-shared-data-access/src/index.d.ts
@@ -15,6 +15,7 @@ export declare const ImportJobStatus: {
   readonly RUNNING: 'RUNNING';
   readonly COMPLETE: 'COMPLETE';
   readonly FAILED: 'FAILED';
+  readonly STOPPED: 'STOPPED';
 };
 
 // packages/spacecat-shared-data-access/src/models/importer/import-constants.js
@@ -393,9 +394,9 @@ export interface Organization {
 export interface Configuration {
   /**
    * Retrieves the configuration version.
-   * @returns {string} The configuration version.
+   * @returns {number} The configuration version.
    */
-  getVersion: () => string;
+  getVersion: () => number;
 
   /**
    * Retrieves the queues configuration.
@@ -854,7 +855,7 @@ export interface DataAccess {
   // configuration functions
   getConfiguration: () => Promise<Readonly<Configuration>>
   getConfigurations: () => Promise<Readonly<Configuration>[]>
-  getConfigurationByVersion: (version: string) => Promise<Readonly<Configuration>>
+  getConfigurationByVersion: (version: number) => Promise<Readonly<Configuration>>
   updateConfiguration: (configurationData: object) => Promise<Readonly<Configuration>>
 
   // key events functions