+++ packages/spacecat-shared-data-access/test/it/generateSampleData.js
@@ -304,7 +304,7 @@ export default async function generateSampleData(
       imports: 'sqs://.../spacecat-services-import-jobs',
       reports: 'sqs://.../spacecat-services-report-jobs',
     },
-    version: 'v2',
+    version: 2,
     PK: config.pkAllConfigurations,
   },
   {
@@ -330,7 +330,7 @@ export default async function generateSampleData(
       audits: 'sqs://.../spacecat-services-audit-jobs',
       reports: 'sqs://.../spacecat-services-report-jobs',
     },
-    version: 'v1',
+    version: 1,
     PK: config.pkAllConfigurations,
   }];