+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,31 @@
+# [@adobe/spacecat-shared-google-client-v1.2.19](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.18...@adobe/spacecat-shared-google-client-v1.2.19) (2024-11-08)
+
+
+### Bug Fixes
+
+* engine spec ([#431](https://github.com/adobe/spacecat-shared/issues/431)) ([f19600e](https://github.com/adobe/spacecat-shared/commit/f19600e74ae7ee42a97c08d5b1c30db79c8ec13d))
+
+# [@adobe/spacecat-shared-google-client-v1.2.18](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.17...@adobe/spacecat-shared-google-client-v1.2.18) (2024-11-07)
+
+
+### Bug Fixes
+
+* set correct node engine spec ([#423](https://github.com/adobe/spacecat-shared/issues/423)) ([5086ce7](https://github.com/adobe/spacecat-shared/commit/5086ce7d41bebb502cadf092e5b9a7b84b5bf103))
+
+# [@adobe/spacecat-shared-google-client-v1.2.17](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.16...@adobe/spacecat-shared-google-client-v1.2.17) (2024-11-02)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#419](https://github.com/adobe/spacecat-shared/issues/419)) ([6ee73f3](https://github.com/adobe/spacecat-shared/commit/6ee73f3aa944bda730276d7a06a588f39b52a56a))
+
+# [@adobe/spacecat-shared-google-client-v1.2.16](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.15...@adobe/spacecat-shared-google-client-v1.2.16) (2024-10-26)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#413](https://github.com/adobe/spacecat-shared/issues/413)) ([ee2c715](https://github.com/adobe/spacecat-shared/commit/ee2c715e08034bea8fb88b4f7166d40b18e107c4))
+
 # [@adobe/spacecat-shared-google-client-v1.2.15](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.14...@adobe/spacecat-shared-google-client-v1.2.15) (2024-10-21)